Fix resume and add progress
Fixed the resume functionality by using an abolute RTP timestamp which
always increases by one. It is not reset on resume anymore. This change
was required in order to allow a correct implementation of the
set_progress function. The TEARDOWN request to disconnect an RTSP
connection is already implemented, but for some reason the connection is
not cleared correctly (maybe there is still a socket open?), which blocks
streaming to the same receiver for a second time, until the receiver is restarted.