Fix RTSP connection and stop function
The RTSP socket is now only opened if we start the handshake and not
when initializing the object. The last behaviour caused an open socket
for each detected device, no matter if it was used or not. The stop
function now resets the progress as aspected and does not resume the
playback. A simpe pyQt5 example was added, which still misses a couple
of features. There is a bug inside the project, which prevents streaming
to multiple audio receivers.