Restructured code
audiosync as well as udpserver now have a reference to all receivers. If
a sync or a resend is required, these two components communicate by
informing the playgroup they belong to, which then instructs the other
component to perform the action. audiosync now uses one global socket,
instead of multiple audiosockets for each device.

There are still mayor problems remaining: No audio packets should be
send in the first two seconds, but a control sync has to be send. This
has to be checked for each device. It's still not possible to stream to
more then one device at a time...