Fix small bugs
Some work on the audio streaming problem was done. You are now able to
stream audio to an AppleTV4. Sometimes it's possible to stream to
multiple devices at once, sometimes not... Shairport-sync still causes a
burst of resend requests. There was a bug inside rtspresponse, which was
fixed by changing the regex. The RTSP timeout was improved, by using the
inbuild timeout function of the queue.