Add initial support for remote cammands
A new subpackage was introduced which starts a http server, that listens
for remote commands (pause / play). While this is working as expected,
Apple TV seems to ignore the protocol and just sends pause and play
randomly... I need to look into this, after I fixed the audio playback /
made it less buggy.